requirements.txt
Django>=3.2,<4.0
django-cors-headers
dj_static
django-environ
djangorestframework
psycopg2-binary
gunicorn